users-service/requirements.txt
fastapi
uvicorn[standard]
motor
python-jose[cryptography]
passlib[bcrypt]
pydantic
